requirements.txt
aiohttp>=3.9.0
httpx>=0.27.0
pyyaml>=6.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
